+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 Changes:
 
   * Allow editing sequencing control types
+  * Indicate library aliquots that are sequencing controls on the Edit Pool page
   * Improved performace of Worksets list page
   * Always show MISO error page instead of default Tomcat error page